# api/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Raikada Recording Server API
  description: |
    HTTP control API exposed by the Raikada Recording Server — the
    on-prem appliance defined in this repository. This is the recorder's
    local control surface. It is invoked by the Management Server over
    the recorder-initiated WebSocket per ADR 0003 (Connectivity model),
    and by LAN clients when authorized. Off-LAN clients reach this
    surface via the Cloud-brokered FRP tunnel.

    **/v1 reshape (ADR 0009).** This document describes the canonical
    `/v1/...` API per ADR 0009. The MediaMTX-lineage `/v3` surface is
    gone; there is no deprecation window and no alias. Every endpoint
    keys on canonical UUIDs (per ADR 0009 §D4); every request and
    response shape uses **snake_case** JSON keys. The single small
    legacy carve-out is `Stream.tracks[].codec_props`: those nine
    discriminated codec sub-shapes inherit the upstream MediaMTX
    camelCase JSON tags (`channelCount`, `sampleRate`, `bitDepth`,
    `muLaw`) and have not been renamed; the inconsistency is tracked
    as a Phase 3 follow-up.

    **In-house error envelope (D15).** Errors return the in-house
    `{ status: "error", error: <message> }` shape, not RFC 7807
    `application/problem+json`. Successful writes that have nothing to
    say return `{ status: "ok" }`. The decision to keep the in-house
    shape is documented in canonical-divergence D15 and a future
...

...

...

...

...
      with them.

...

...

...

...

...

...

...
servers:
...
security:
  - basicAuth: []
  - bearerJWT: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        HTTP Basic credentials. Validated against the recorder's
        configured auth manager (internal, HTTP, or JWT-backed) on
        every request.
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Standard JWT bearer token. Validated against JWKS held by the
        recorder's auth manager. Use
        `POST /v1/auth/refresh-issuer-material` to force a JWKS
        reload after key rotation.

        Note: a `Bearer` token that happens to contain a colon (e.g.
        the upstream MediaMTX `user:pass` convention) is treated
        opaquely as a token here, not as colon-separated credentials.
        Clients that need username+password authentication must use
        `basicAuth`. See D6 in
        `recorder/docs/canonical-divergences.md`.
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
      description: Zero-based page index.
    ItemsPerPage:
      name: items_per_page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
      description: |
        Items per page; default 100. The list-response wrapper
        carries `item_count` (total before pagination), `page_count`,
        and the windowed `items` array.
  schemas:
...
    OKStatus:
      type: string
      enum: [ok]
    ErrorStatus:
      type: string
      enum: [error]
    OK:
      type: object
      description: |
        Successful-write envelope. Used for `DELETE` and PATCH calls
        that have no body to return. Per canonical-divergence D15
        this is an in-house shape, not RFC 7807.
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/OKStatus'
    Error:
      type: object
      description: |
        In-house error envelope. The HTTP status code carries the
        primary signal; `error` is a free-form human-readable
        message. Per canonical-divergence D15 a future
        platform-API-contracts ADR may migrate this to RFC 7807.
      required: [status, error]
      properties:
        status:
          $ref: '#/components/schemas/ErrorStatus'
        error:
          type: string
    PaginationListWrapper:
      type: object
      description: |
        Common pagination envelope. Concrete list responses substitute
        their own item type for `items[]`.
      required: [item_count, page_count, items]
      properties:
        item_count:
          type: integer
          description: Total items matching the query, before pagination.
        page_count:
          type: integer
          description: Total page count for the current `items_per_page`.
        items:
          type: array
          items: {}

    # ----- Info -----------------------------------------------------
...
    StreamTrackKind:
      type: string
      enum:
        - video
        - audio
        - metadata
...
    RecordingState:
      type: string
      description: |
        Recording lifecycle: `active → sealed`; `sealed → archived |
        deleted`; `(any) → failed`.
      enum:
        - active
        - sealed
        - archived
        - failed
        - deleted
    RecordingSegmentContainer:
      type: string
      enum:
        - fmp4
        - mpegts
    RecordingSegmentState:
      type: string
      enum:
        - recording
        - sealed
        - archived
        - deleted
...
    RecordingSegmentTrack:
      type: object
      required: [kind, codec]
      properties:
        kind:
          $ref: '#/components/schemas/StreamTrackKind'
        codec:
          type: string
    RecordingSegment:
      type: object
      description: |
        Canonical 17-field RecordingSegment entity. Adds
        `recording_id` per ADR 0009. The recorder-local on-disk
        `path` field documented in `domain-model.md` is excluded
        from API responses by classification rule. Closes
        canonical-divergence D8.
      required:
        - id
        - tenant_id
        - site_id
        - recording_server_id
        - camera_id
        - volume_id
        - policy_id
        - recording_id
        - started_at
        - ended_at
        - duration
        - container
        - size_bytes
        - state
        - created_at
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        site_id:
          type: string
          format: uuid
        recording_server_id:
          type: string
          format: uuid
        camera_id:
          type: string
          format: uuid
        volume_id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
        recording_id:
          type: string
          format: uuid
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        duration:
          type: string
        container:
          $ref: '#/components/schemas/RecordingSegmentContainer'
        size_bytes:
          type: integer
          format: int64
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/RecordingSegmentTrack'
        checksum:
          type: string
        state:
          $ref: '#/components/schemas/RecordingSegmentState'
        created_at:
          type: string
          format: date-time
...
    Recording:
      type: object
      description: |
        Canonical Recording entity introduced inline by ADR 0009. A
        continuous span of recorded footage on one camera, composed
        of one or more `RecordingSegment`s. The client/scrub
        primitive. List responses omit `segments`; GET-by-id
        includes them.
      required:
        - id
        - tenant_id
        - site_id
        - camera_id
        - recording_server_id
        - started_at
        - state
        - duration
        - byte_size
        - segment_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        site_id:
          type: string
          format: uuid
        camera_id:
          type: string
          format: uuid
        recording_server_id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        state:
          $ref: '#/components/schemas/RecordingState'
        failure_reason:
          type: string
          description: |
            Required when `state == failed`; null otherwise.
        duration:
          type: string
        byte_size:
          type: integer
          format: int64
        segment_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        segments:
          type: array
          description: |
            Embedded only on the GET-by-id response per ADR 0009 §D5;
            the list endpoint omits this field.
          items:
            $ref: '#/components/schemas/RecordingSegment'
    RecordingList:
      allOf:
        - $ref: '#/components/schemas/PaginationListWrapper'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Recording'
...
paths:
...
  /v1/recordings:
    get:
      tags: [Recordings]
      summary: List recordings.
      description: |
        Returns canonical Recording entities. List responses **omit**
        the embedded `segments` array per ADR 0009 §D5 — fetch a
        single recording for the segment list.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/ItemsPerPage'
        - name: camera_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: state
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RecordingState'
        - name: started_after
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: started_before
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Recording list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingList'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/recordings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Recordings]
      summary: Get a recording with embedded segments.
      description: |
        Embeds the segment list per ADR 0009 §D5. Each segment has
        its recorder-local `path` field scrubbed before serialization.
      responses:
        '200':
          description: Recording detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recording'
        '400':
          description: Invalid recording id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Recording not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Recordings]
      summary: Delete a recording and all of its segments.
      description: |
        Cascades through every `RecordingSegment` belonging to the
        Recording, removing each file from disk. Equivalent to a
        client iterating `DELETE /v1/recording-segments/{id}` for each
        segment, but issued in a single call so clients don't race
        against the recorder rotating new segments mid-iteration on
        an active Recording.

        Active Recordings are deletable; the in-flight segment is
        removed along with the rest. The recorder's record loop will
        discover the missing file on its next write and the
        `segment.write_failed` Event surfaces the condition.

        Per ADR 0009 §D5 the Recording's canonical lifecycle is to
        transition to `state=deleted` and persist as a tombstone for
        audit; the recorder's current synthesis-from-disk model has
        no persistent Recording row, so the post-DELETE GET returns
        404 instead. Persistent Recording storage is its own
        follow-up; the audit-chain entry written by this handler is
        the durable record today.

        Partial failure (one or more segments couldn't be removed
        for a non-`ENOENT` reason) returns 500 with a per-segment
        error list; the cache is invalidated regardless so a follow-
        up GET reflects whatever survived.
      responses:
        '200':
          description: Recording (and all segments) deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '400':
          description: Invalid recording id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Recording not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Partial failure during cascade; per-segment errors in body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...